Rename generated.py to key.py
